rust cli: rename add/amend.rs to add/shared.rs
amend.rs sat alongside attachment.rs and metadata.rs (the real
add subcommands), so it read like a third subcommand. Rename to
shared.rs to signal it is the support module shared by the add
subcommands, not a subcommand itself. Rename only.
