+++ rust/cli/src/commands/add.rs
@@ -1,5 +1,5 @@
 //! `mcap add` subcommands and the shared in-place amendment engine.
 
-pub(crate) mod amend;
 pub(crate) mod attachment;
 pub(crate) mod metadata;
+pub(crate) mod shared;
