+++ rust/cli/src/commands/add/shared.rs
